pyproject.toml
[project]
name = "jominju-week9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "numpy>=2.3.5",
    "pygame>=2.6.1",
]
